NamedParameterJdbcTemplate 사용

+++ src/main/java/com/weblibrary/domain/user/repository/JdbcUserRepository.java
@@ -1,30 +1,31 @@
 package com.weblibrary.domain.user.repository;
 
-import com.weblibrary.domain.user.exception.NotFoundUserException;
 import com.weblibrary.domain.user.model.User;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 import javax.sql.DataSource;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Repository
 public class JdbcUserRepository implements UserRepository {
 
-    private final JdbcTemplate template;
+    private final NamedParameterJdbcTemplate template;
     private final SimpleJdbcInsert jdbcInsert;
 
     public JdbcUserRepository(DataSource dataSource) {
-        this.template = new JdbcTemplate(dataSource);
+        this.template = new NamedParameterJdbcTemplate(dataSource);
         this.jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName("users")
                 .usingGeneratedKeyColumns("user_id");
@@ -40,9 +41,10 @@ public class JdbcUserRepository implements UserRepository {
 
     @Override
     public Optional<User> findById(Long id) {
-        String sql = "select * from users where user_id = ?";
+        String sql = "select * from users where user_id = :userId";
         try {
-            User user = template.queryForObject(sql, getUserRowMapper(), id);
+            Map<String, Long> param = Map.of("userId", id);
+            User user = template.queryForObject(sql, param, getUserRowMapper());
             return Optional.ofNullable(user);
         } catch (DataAccessException e) {
             return Optional.empty();
@@ -51,9 +53,10 @@ public class JdbcUserRepository implements UserRepository {
 
     @Override
     public Optional<User> findByUsername(String username) {
-        String sql = "select * from users where username = ?";
+        String sql = "select * from users where username = :username";
         try {
-            User user = template.queryForObject(sql, getUserRowMapper(), username);
+            Map<String, String> param = Map.of("username", username);
+            User user = template.queryForObject(sql, param, getUserRowMapper());
             return Optional.ofNullable(user);
         } catch (DataAccessException e) {
             return Optional.empty();
@@ -62,10 +65,13 @@ public class JdbcUserRepository implements UserRepository {
 
     @Override
     public Page<User> findAll(Pageable pageable) {
-        String sql = "select * from users order by user_id desc limit ? offset ?";
+        String sql = "select * from users order by user_id desc limit :limit offset :offset";
         List<User> users = null;
         try {
-            users = template.query(sql, getUserRowMapper(), pageable.getPageSize(), pageable.getOffset());
+            SqlParameterSource param = new MapSqlParameterSource()
+                    .addValue("limit", pageable.getPageSize())
+                    .addValue("offset", pageable.getOffset());
+            users = template.query(sql, param, getUserRowMapper());
         } catch (DataAccessException e) {
             users = List.of();
         }
@@ -74,7 +80,7 @@ public class JdbcUserRepository implements UserRepository {
         String countSql = "select count(*) from users";
         int total = 0;
         try {
-            total = template.queryForObject(countSql, Integer.class);
+            total = template.queryForObject(countSql, Map.of(), Integer.class);
         } catch (DataAccessException e) {
             total = 0;
         }
@@ -83,30 +89,22 @@ public class JdbcUserRepository implements UserRepository {
     }
 
     @Override
-    public User update(User user) {
-        User oldUser = findById(user.getUserId())
-                .orElseThrow(NotFoundUserException::new);
-
+    public void update(User user) {
+        SqlParameterSource param = new BeanPropertySqlParameterSource(user);
         String sql = "update users set " +
-                "username = ?, " +
-                "password = ?, " +
-                "remaining_rents = ? " +
-                "where user_id = ?";
+                "username = :username, " +
+                "password = :password, " +
+                "remaining_rents = :remainingRents " +
+                "where user_id = :userId";
 
-        template.update(sql, user.getUsername(), user.getPassword(), user.getRemainingRents(), user.getUserId());
-
-        return oldUser;
+        template.update(sql, param);
     }
 
     @Override
-    public Optional<User> remove(Long userId) {
-        String sql = "delete from users where user_id=?";
-
-        User user = findById(userId)
-                .orElseThrow(NotFoundUserException::new);
-
-        template.update(sql, userId);
-        return Optional.of(user);
+    public void remove(Long userId) {
+        String sql = "delete from users where user_id=:userId";
+        Map<String, Long> param = Map.of("userId", userId);
+        template.update(sql, param);
     }
